+++ internal/models/character.go
@@ -1,8 +1,8 @@
 package models
 
 type Character struct {
-	ID         int    `db:"id" json:"id"`
-	UserID     int    `db:"user_id" json:"user_id"`
+	ID         int64  `db:"id" json:"id"`
+	UserID     int64  `db:"user_id" json:"user_id"`
 	Name       string `db:"name" json:"name"`
 	Gender     string `db:"gender" json:"gender"`
 	AvatarID   string `db:"avatar_id" json:"avatar_id"`
