+++ internal/models/refresh_token.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type RefreshToken struct {
-	ID        int       `db:"id"`
-	UserID    int       `db:"user_id"`
+	ID        int64     `db:"id"`
+	UserID    int64     `db:"user_id"`
 	Token     string    `db:"token"`
 	ExpiresAt time.Time `db:"expires_at"`
 	CreatedAt time.Time `db:"created_at"`
