+++ internal/models/user.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type User struct {
-	ID        int       `db:"id" json:"id"`
+	ID        int64     `db:"id" json:"id"`
 	Username  string    `db:"username" json:"username"`
 	Email     string    `db:"email" json:"email"`
 	Password  string    `db:"password" json:"password,omitempty"`
